Move pending transactions logic to different file SDKv5
