+++ services/core/shared/core/compat/sdk_v5/transactions.js
@@ -13,10 +13,7 @@
  * Removal or modification of this copyright notice is prohibited.
  *
  */
-const { Logger } = require('lisk-service-framework');
 const BluebirdPromise = require('bluebird');
-
-const logger = Logger();
 
 const coreApi = require('./coreApi');
 const {
@@ -33,7 +30,6 @@ const transactionsIndexSchema = require('./schema/transactions');
 const getTransactionsIndex = () => mysqlIndex('transactions', transactionsIndexSchema);
 
 const availableLiskModuleAssets = getRegisteredModuleAssets();
-let pendingTransactionsList = [];
 
 const resolveModuleAsset = (moduleAssetVal) => {
 	const [module, asset] = moduleAssetVal.split(':');
@@ -202,109 +198,7 @@ const getTransactions = async params => {
 	return transactions;
 };
 
-const getPendingTransactionsFromCore = async () => {
-	const response = await coreApi.getPendingTransactions();
-	let pendingTx = response.data.map(tx => normalizeTransaction(tx));
-	pendingTx = await BluebirdPromise.map(
-		pendingTx,
-		async transaction => {
-			const account = await getIndexedAccountInfo({ publicKey: transaction.senderPublicKey });
-			transaction.senderId = account && account.address ? account.address : undefined;
-			transaction.username = account && account.username ? account.username : undefined;
-			transaction.isPending = true;
-			return transaction;
-		},
-		{ concurrency: pendingTx.length },
-	);
-	return pendingTx;
-};
-
-const loadAllPendingTransactions = async () => {
-	pendingTransactionsList = await getPendingTransactionsFromCore();
-	logger.info(`Initialized/Updated pending transactions cache with ${pendingTransactionsList.length} transactions.`);
-};
-
-const getPendingTransactions = async params => {
-	const pendingTransactions = {
-		data: [],
-		meta: {},
-	};
-	const requestParams = {};
-	const offset = Number(params.offset) || 0;
-	const limit = Number(params.limit) || 10;
-
-	if (params.sort && params.sort.includes('nonce') && !params.senderId) {
-		throw new Error('Nonce based sorting is only possible along with senderId');
-	}
-
-	if (params.username) {
-		const [accountInfo] = await getIndexedAccountInfo({ username: params.username });
-		if (!accountInfo || accountInfo.address === undefined) return new Error(`Account with username: ${params.username} does not exist`);
-		requestParams.senderPublicKey = accountInfo.publicKey;
-	}
-
-	if (params.senderIdOrRecipientId) {
-		requestParams.senderId = params.senderIdOrRecipientId;
-		requestParams.recipientId = params.senderIdOrRecipientId;
-	}
-
-	if (params.senderId) {
-		const account = await getIndexedAccountInfo({ address: params.senderId });
-		requestParams.senderPublicKey = account.publicKey;
-	}
-
-	if (params.amount && params.amount.includes(':')) {
-		const minAmount = params.amount.split(':')[0];
-		const maxAmount = params.amount.split(':')[1];
-		requestParams.minAmount = Number(minAmount);
-		requestParams.maxAmount = Number(maxAmount);
-	}
-
-	const sortComparator = (sortParam) => {
-		const sortProp = sortParam.split(':')[0];
-		const sortOrder = sortParam.split(':')[1];
-
-		const comparator = (a, b) => (sortOrder === 'asc')
-			? Number(a[sortProp]) - Number(b[sortProp])
-			: Number(b[sortProp]) - Number(a[sortProp]);
-		return comparator;
-	};
-
-	if (pendingTransactionsList.length) {
-		const filteredPendingTxs = pendingTransactionsList.filter(transaction => (
-			(!requestParams.senderPublicKey
-				|| transaction.senderPublicKey === requestParams.senderPublicKey)
-			&& (!requestParams.recipientId
-				|| transaction.asset.recipientAddress === requestParams.recipientId)
-			&& (!requestParams.moduleAssetId
-				|| transaction.amoduleAssetId === requestParams.moduleAssetId)
-			&& (!requestParams.moduleAssetName
-				|| transaction.moduleAssetName === requestParams.moduleAssetName)
-			&& (!requestParams.data
-				|| transaction.asset.data.includes(requestParams.data))
-			&& (!requestParams.data
-				|| transaction.asset.data.includes(requestParams.data))
-			&& (!requestParams.from
-				|| Number(transaction.amount) >= requestParams.minAmount)
-			&& (!requestParams.to
-				|| Number(transaction.amount) <= requestParams.maxAmount)
-		));
-		pendingTransactions.data = filteredPendingTxs
-			.sort(sortComparator(params.sort))
-			.slice(offset, offset + limit);
-
-		pendingTransactions.meta = {
-			count: pendingTransactions.data.length,
-			offset,
-			total: filteredPendingTxs.length,
-		};
-	}
-	return pendingTransactions;
-};
-
 module.exports = {
 	getTransactions,
 	indexTransactions,
-	getPendingTransactions,
-	loadAllPendingTransactions,
 };
